minor: 调整单元测试
