+++ workloads/paas_wl/tests/networking/ingress/domains/test_independent.py
@@ -33,6 +33,7 @@ class TestReplaceAppDomainService:
         with pytest.raises(ReplaceAppDomainFailed):
             ReplaceAppDomainService(bk_stag_env, 'invalid-name.example.com', '/')
 
+    @pytest.mark.mock_get_structured_app
     @pytest.mark.auto_create_ns
     @pytest.mark.parametrize(
         'old_host,old_path_prefix,new_host,new_path_prefix',
